# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduMaster 360 - School Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /students/{id}/profile:
    get:
      operationId: getStudentProfile
      tags: [students]
      summary: Get full 360 student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

  # ─── TEACHERS ────────────────────────────────────────────────────────
  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List all teachers
      responses:
        "200":
          description: Teachers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{id}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get teacher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    patch:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"

  # ─── CLASSES ─────────────────────────────────────────────────────────
...
  /suggestions:
    get:
      operationId: listSuggestions
      tags: [suggestions]
      summary: List suggestions (admin/headteacher only)
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Suggestion"
    post:
      operationId: createSuggestion
      tags: [suggestions]
      summary: Submit a suggestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suggestion"
  /suggestions/{id}:
    get:
      operationId: getSuggestion
      tags: [suggestions]
      summary: Get suggestion detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suggestion"
    patch:
      operationId: updateSuggestion
      tags: [suggestions]
      summary: Update suggestion status / add admin response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Suggestion"

  # ─── ANALYTICS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dateOfBirth:
          type: string
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        gender:
          type: string
          enum: [male, female, other]
        photoUrl:
          type: ["string", "null"]
        parentContact:
          type: ["string", "null"]
        performanceScore:
          type: ["number", "null"]
        riskScore:
          type: ["number", "null"]
        riskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        createdAt:
          type: string
      required: [id, name, dateOfBirth, gender, createdAt]
    StudentInput:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
        classId:
          type: ["integer", "null"]
        gender:
          type: string
          enum: [male, female, other]
        photoUrl:
          type: ["string", "null"]
        parentContact:
          type: ["string", "null"]
      required: [name, dateOfBirth, gender]
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
        classId:
          type: ["integer", "null"]
        gender:
          type: string
          enum: [male, female, other]
        photoUrl:
          type: ["string", "null"]
        parentContact:
          type: ["string", "null"]
    StudentProfile:
      type: object
      properties:
        student:
          $ref: "#/components/schemas/Student"
        marks:
          type: array
          items:
            $ref: "#/components/schemas/Mark"
        attendance:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        disciplineRecords:
          type: array
          items:
            $ref: "#/components/schemas/DisciplineRecord"
        fees:
          type: array
          items:
            $ref: "#/components/schemas/FeeRecord"
        ranking:
          $ref: "#/components/schemas/StudentRanking"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
      required: [student, marks, attendance, disciplineRecords, fees]
    TimelineEvent:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [mark, attendance, discipline, fee, announcement]
        title:
          type: string
        description:
          type: string
        date:
          type: string
        severity:
          type: ["string", "null"]
          enum: [positive, neutral, negative, null]
      required: [id, type, title, description, date]

    # ─── TEACHERS ────────────────────────────────────────────────────
    Teacher:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        photoUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        classIds:
          type: array
          items:
            type: integer
        performanceScore:
          type: ["number", "null"]
        attendanceRate:
          type: ["number", "null"]
        marksSubmitted:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, name, subject, createdAt]
    TeacherInput:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        photoUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
      required: [name, subject]
    TeacherUpdate:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        photoUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    # ─── CLASSES ──────────────────────────────────────────────────────
...
    Mark:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        subject:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: ["number", "null"]
        grade:
          type: ["string", "null"]
        term:
          type: string
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, subject, score, maxScore, term, createdAt]
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        date:
          type: string
        status:
          type: string
          enum: [present, absent, late, excused]
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, date, status, createdAt]
...
    FeeRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        term:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        status:
          type: string
          enum: [paid, partial, unpaid]
        bursarId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, term, totalAmount, amountPaid, balance, status, createdAt]
...
    DisciplineRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [warning, suspension, commendation, detention, expulsion]
        description:
          type: string
        date:
          type: string
        issuedBy:
          type: ["string", "null"]
        severity:
          type: string
          enum: [low, medium, high]
        resolved:
          type: boolean
        createdAt:
          type: string
      required: [id, studentId, type, description, date, severity, resolved, createdAt]
...
    Suggestion:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: ["integer", "null"]
        senderName:
          type: ["string", "null"]
        senderRole:
          type: ["string", "null"]
        message:
          type: string
        category:
          type: string
          enum: [academic, facilities, discipline, welfare, management, other]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, reviewed, resolved]
        adminResponse:
          type: ["string", "null"]
        anonymous:
          type: boolean
        createdAt:
          type: string
      required: [id, message, category, priority, status, anonymous, createdAt]
    SuggestionInput:
      type: object
      properties:
        message:
          type: string
        category:
          type: string
          enum: [academic, facilities, discipline, welfare, management, other]
        anonymous:
          type: boolean
      required: [message, category]
    SuggestionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewed, resolved]
        priority:
          type: string
          enum: [low, medium, high]
        adminResponse:
          type: ["string", "null"]

    # ─── ANALYTICS ───────────────────────────────────────────────────
...
    StudentRanking:
      type: object
      properties:
        studentId:
          type: integer
        name:
          type: string
        className:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        overallScore:
          type: number
        academicScore:
          type: number
        disciplineScore:
          type: number
        attendanceScore:
          type: number
        rank:
          type: integer
        classRank:
          type: ["integer", "null"]
        trend:
          type: string
          enum: [up, down, stable]
      required: [studentId, name, overallScore, academicScore, disciplineScore, attendanceScore, rank, trend]
...